media-source/MediaSource-addSourceBuffer_readyState.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>readyState attribute not in the "open" state</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = 'test.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';
async_test(function(t) {
    var mediaSource = new MediaSource();
    t.step(function() {
        assert_throws(
            {name : 'InvalidStateError'},
            function() {
                var sourceBuffer = mediaSource.addSourceBuffer(mime);
            },
            'Set invalid duration'
        );
    });
    t.done();
}, 'MediaSource#addSourceBuffer : If the readyState attribute is not in the "open" state then throw an INVALID_STATE_ERR exception and abort these steps.');
</script>
</body>
</html>
